3 commits in one, had to reset them because one of them contained wrong files. Add: App remembers task you created previously. Note: it still doesnt remember the changes you make, only creation params. Fix: not showing task info when clicked on mini widget Update: merged files: tasks+tasksteps+toggle -> widgets, ui_constants+time_constants -> constants. Delete: load_ui.py

+++ data/init_db.py
@@ -1,12 +1,22 @@
-import sqlite3
 from datetime import datetime, timezone
+import sqlite3
+import os
 
-conn = sqlite3.connect("data/user_data.db")
+from config.env_loader import drop_db_mode
+
+BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+db_path = os.path.join(BASE_DIR, 'user_data.db')
+
+conn = sqlite3.connect(db_path)
+
 cursor = conn.cursor()
 
-cursor.execute("""
-DROP TABLE IF EXISTS users 
-""")
+#for test cases
+if drop_db_mode:
+    cursor.execute("""
+    DROP TABLE IF EXISTS users
+    """)
+
 cursor.execute("""
 CREATE TABLE IF NOT EXISTS users (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -21,11 +31,10 @@ CREATE TABLE IF NOT EXISTS users (
     repeatable INTEGER,
     rep_option TEXT,
     rep_vals TEXT,
-    task_steps_ammo INTEGER,
     task_steps_infos TEXT, 
-    
+    description TEXT
 )
 """)
 
 conn.commit()
-conn.close()
+#conn.close()
